requirements.txt
Flask==2.3.3
numpy==1.24.3
Pillow==9.5.0
pydub==0.25.1
requests==2.31.0
opencv-python==4.8.0.74
